+++ src/bittytax/conv/parsers/celsius.py
@@ -21,8 +21,6 @@ def parse_celsius(data_row, parser, **kwargs):
     if row_dict["Confirmed"] != "Yes" and not kwargs["unconfirmed"]:
         sys.stderr.write(
             f"{Fore.YELLOW}row[{parser.in_header_row_num + data_row.line_num}] {data_row}\n"
-        )
-        sys.stderr.write(
             f"{WARNING} Skipping unconfirmed transaction, use the [-uc] option to include it\n"
         )
         return